requirements.txt
Flask==2.2.2
pymongo==4.2.0
dnspython==2.2.1
gunicorn==20.1.0
python-dotenv==0.20.0